let's try this one...

+++ moneth/main.cpp
@@ -102,7 +102,7 @@ h256 fromHexToH256(const string& s)
 {
 	h256 result;
 	bytes b = fromHex(s);
-	memcpy(result.data(), b.data(), std::min<uint>(b.size(), (uint)result.size));
+	memcpy(result.data(), b.data(), std::min<size_t>(b.size(), (size_t)result.size));
 	return result;
 }
